requirements_multimodal.txt
# Multimodal dependencies for VHAS (no Kaggle required)

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Image/video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Pose estimation (optional, but recommended)
mediapipe>=0.10.0

# Numerical
numpy>=1.21.0